refactoring som e test to its right place
refactoring som e test to its right place
