+++ module/src/test/java/eu/europa/ec/fisheries/uvms/rest/asset/service/AssetRestResourceTest.java
@@ -1241,21 +1241,4 @@ public class AssetRestResourceTest extends AbstractAssetRestTest {
         }
         return mtList;
     }
-
-    @Test
-    @OperateOnDeployment("normal")
-    public void changeHistoryRowTest () {
-        ChangeHistoryRow changeHistoryRow = new ChangeHistoryRow("updated", Instant.now());
-        assertEquals("updated", changeHistoryRow.getUpdatedBy());
-        assertNotNull(changeHistoryRow.getUpdateTime());
-    }
-
-    @Test
-    @OperateOnDeployment("normal")
-    public void changeHistoryItemTest () {
-        ChangeHistoryItem changeHistoryItem = new ChangeHistoryItem("field", "oldvalue", "newvalue");
-        assertEquals("field", changeHistoryItem.getField());
-        assertEquals("oldvalue", changeHistoryItem.getOldValue());
-        assertEquals("newvalue", changeHistoryItem.getNewValue());
-    }
 }
